pull methods and definitions into their own files for easier review
